view previously completed

+++ src/app/components/Todo.tsx
@@ -5,12 +5,13 @@ import InlineEdit from './InlineEdit';
 interface Props {
     data: TodoItem
     className?: string
-    onDoneToggle: (todo: TodoItem, done: boolean) => void
-    onEditClicked: () => void
-    onEdited: (prevTodo: TodoItem, updatedTodo: TodoItem) => void
-    onEditCancel: () => void
-    onDelete: (todo: TodoItem) => void
-    edit: boolean
+    onDoneToggle?: (todo: TodoItem, done: boolean) => void
+    onEditClicked?: () => void
+    onEdited?: (prevTodo: TodoItem, updatedTodo: TodoItem) => void
+    onEditCancel?: () => void
+    onDelete?: (todo: TodoItem) => void
+    edit?: boolean
+    readonly?: boolean
 }
 
 interface State {
@@ -30,10 +31,11 @@ export default class Todo extends React.PureComponent<Props, State> {
         this.editItem = this.editItem.bind(this)
         this.onEditComplete = this.onEditComplete.bind(this)
         this.deleteItem = this.deleteItem.bind(this)
+        this.onToggleDone = this.onToggleDone.bind(this)
     }
 
     private onHover() {
-        if (!this.state.hover) {
+        if (!this.state.hover && !this.props.readonly) {
             this.setState({hover: true})
         }
     }
@@ -44,24 +46,37 @@ export default class Todo extends React.PureComponent<Props, State> {
 
     private editItem(e: React.MouseEvent<HTMLAnchorElement>) {
         e.stopPropagation()
-        this.props.onEditClicked()
+        if (this.props.onEditClicked && !this.props.readonly) {
+            this.props.onEditClicked()
+        }
     }
+
     private onEditComplete(data: {value: string, wasEnter: boolean}) {
-        this.props.onEdited(this.props.data, Object.assign({}, this.props.data, {description: data.value}))
+        if (this.props.onEdited && !this.props.readonly) {
+            this.props.onEdited(this.props.data, Object.assign({}, this.props.data, {description: data.value}))
+        }
     }
 
     private deleteItem(e: React.MouseEvent<HTMLAnchorElement>) {
         e.stopPropagation()
-        this.props.onDelete(this.props.data)
+        if (this.props.onDelete && !this.props.readonly) {
+            this.props.onDelete(this.props.data)
+        }
+    }
+
+    private onToggleDone() {
+        if (this.props.onDoneToggle && !this.props.readonly) {
+            this.props.onDoneToggle(this.props.data, !this.props.data.done)
+        }
     }
 
     render() {
         let todoData = <>{this.props.data.description}</>
 
         if (this.props.edit) {
-            todoData = <InlineEdit editing={true} onCancel={this.props.onEditCancel} className="row is-full" value={this.props.data.description} onComplete={this.onEditComplete} />
+            todoData = <InlineEdit editing={true} onCancel={this.props.onEditCancel!} className="row is-full" value={this.props.data.description} onComplete={this.onEditComplete} />
         }
-        return <div className={`row todo ${this.props.className} ${this.props.data.done ? 'done' : ''}`} onMouseOver={this.onHover} onMouseOut={this.onHoverStop} onClick={() => this.props.onDoneToggle(this.props.data, !this.props.data.done)}>
+        return <div className={`row todo ${this.props.className} ${this.props.data.done && !this.props.readonly ? 'done' : ''} ${this.props.readonly ? 'readonly' : ''}`} onMouseOver={this.onHover} onMouseOut={this.onHoverStop} onClick={this.onToggleDone}>
             <div className="handle">{this.state.hover ? '||' : ''}</div>
             <div className="row is-full todo-description">
                 {todoData}
